Data.These.Lens

Traversals

here :: Traversal (These a c) (These b c) a b

A Traversal of the first half of a These, suitable for use with Control.Lens.

>>> over here show (That 1)
That 1

>>> over here show (These 'a' 2)
These "'a'" 2

there :: Traversal (These c a) (These c b) a b

A Traversal of the second half of a These, suitable for use with Control.Lens.

there :: Traversal (These t b) (These t b) a b

>>> over there show (That 1)
That "1"

>>> over there show (These 'a' 2)
These 'a' "2"

Prisms

_This :: Prism' (These a b) a

A Prism' selecting the This constructor.

Note: cannot change type.

_That :: Prism' (These a b) b

A Prism' selecting the That constructor.

Note: cannot change type.

_These :: Prism' (These a b) (a, b)

A Prism' selecting the These constructor. These names are ridiculous!

Note: cannot change type.